// models/ui_homepage_tile_callout.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery'

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type UiHomepageTileCallout = IContentItem<{
	arrow_color: Elements.CustomElement
	divider_color: Elements.CustomElement
	tiles: Elements.LinkedItemsElement<IContentItem>
	font_color: Elements.CustomElement
	background_color: Elements.CustomElement
}>